Split Calculator view into Screen and Keyboard
